requirements.txt
pymupdf>=1.24.0
pyyaml>=6.0
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0
